refactored useGeojsonState, unmemoized SourceGeojson
