+++ app/src/app/regionen/[regionSlug]/_components/Map/SourcesAndLayers/SourceGeojson/SourceGeojson.tsx
@@ -1,14 +1,13 @@
-import React, { memo } from 'react'
+import React from 'react'
 import { Layer, Source } from 'react-map-gl/maplibre'
-import { Store, useGeojsonStore } from './useGeojsonStore'
-
-type Props = Pick<Store, 'geojson'>
+import { useGeojsonData } from './useGeojsonState'
 
 // very useful for debugging geometry:
-// const { setGeojson, resetGeojson } = useGeojsonStore()
+// const { setGeojson, resetGeojson } = useGeojsonActions()
 // setGeojson(geojsonData)
-const SourceGeojsonMemoized = memo(function SourceGeojsonMemoized(props: Props) {
-  const { geojson } = props
+
+export const SourceGeojson = () => {
+  const geojson = useGeojsonData()
   if (!geojson) return null
 
   return (
@@ -24,10 +23,4 @@ const SourceGeojsonMemoized = memo(function SourceGeojsonMemoized(props: Props) 
       />
     </Source>
   )
-})
-
-export const SourceGeojson: React.FC = () => {
-  const { geojson } = useGeojsonStore()
-  const props = { geojson }
-  return <SourceGeojsonMemoized {...props} />
 }
